Migrated pie 3d to class pattern.
