+++ js/masters/highcharts-3d.src.js
@@ -16,5 +16,5 @@ import '../Core/Axis/Axis3D.js';
 import '../Core/Axis/Tick3D.js';
 import '../Core/Series/Series3D.js';
 import '../Series/Column3DSeries.js';
-import '../Series/Pie3DSeries.js';
+import '../Series/Pie3D/Pie3DSeries.js';
 import '../Series/Scatter3DSeries.js';
